+++ test/src/mbtest_test.c
@@ -236,7 +236,7 @@ static int test_coil_read_fn(void)
 	return 1;
 }
 
-static enum mbstatus_e test_coil_write_fn(uint8_t value)
+static enum mbstatus_e test_coil_write_fn(int value)
 {
 	(void)value; /* Suppress unused parameter warning */
 	return MB_OK;
